ui/requirements.txt
pillow
tkcalendar
